pyproject.toml
[project]
name = "junrag"
version = "0.1.0"
description = "A RAG library with query decomposition, late chunking, and contextual retrieval support"
readme = "README.md"
requires-python = ">=3.10,<3.14"
license = { text = "MIT" }
authors = [{ name = "Jun Kim" }]
keywords = ["rag", "retrieval", "llm", "embeddings", "reranking", "query-decomposition"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fire>=0.5.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "tokenizers>=0.15.0",
    "vllm>=0.6.0",
    "qdrant-client>=1.7.0",
    "openai>=1.0.0",
    "tqdm>=4.65.0",
    "datasets>=4.4.1",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]
tools = [
    "requests>=2.28.0",
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.11.0",
]

[project.scripts]
junrag = "junrag.cli.pipeline:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/junrag"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
